// src/lib/api/models/updateDealDtoStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type UpdateDealDtoStatus = typeof UpdateDealDtoStatus[keyof typeof UpdateDealDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateDealDtoStatus = {
  SOFT_RESERVED: 'SOFT_RESERVED',
  NEGOTIATING: 'NEGOTIATING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
  DISPUTED: 'DISPUTED',
} as const;